Add memory checking scripts and deal with TCLAP leak errors.
A script, runValgrind.sh, is added to do local memory checking.  See the
comments in the script or the output of --help for usage.

We ignore the Slicer/TCLAP leak errors with a CTestCustom.cmake file.  These
errors only seem to occur when --help is passed.  The Slicer code uses long
preprocessor definitions, and it is difficult to see what the real problem is,
so the errors are ignored for now.