Start using smart pointers.
Use smart pointers for VectorArray, VectorImage, and VectorField, CState.

A new base class is created for classes that want to have smart pointers,
CObject.  The implementation is copied from ITK so that the itk::SmartPointer
can be used to hold the reference count.  Note that this is lighter weight than
the itk::LightObject from which it is derived -- it only has content for
handling m_ReferenceCount.

The output setting for doxygen were tweaked to be more information to help with
the migration.

There are still some memory leaks to be addressed with future commits.

Change-Id: Ie27e68f87249228a431dd84ee6e905900824b73c