Add CAlgorithmFactory.
Remove lower level operations of reading and writing configuration files from
CAlgorithmBase.  Also, greatly simplify CAlgorithmBase/CAlgorithm be removing
the extra CJSONConfiguration's.  CJSONConfiguration is a made into a smart
pointer.

CAlgorithmFactory gets used in CApplication.

Some basic parsing of a data configuration for the input and outputs is
performed.  Refinement and movement to the ImageManager in the future is likely.

More tests for other algorithms and refinement of the tests to come.

Change-Id: I53f7c7e669a7a9b6bb2eea7c33c4534845bad341